tweak runner to run on macos
Only adapter tested
