+++ adapter/runners/src/sys/linux.rs
@@ -1,11 +1,10 @@
-use nix::sys::stat;
 use nix::unistd::{self, Gid, Uid};
 use users::get_user_by_name;
 
 use std::net::IpAddr;
-use std::os::unix::process::CommandExt;
 use std::process::Command;
 
+use crate::sys::unix as common;
 use crate::sys::{Platform, PlatformErr};
 
 const DEFAULT_TUN_NAME: &str = "tun0";
@@ -14,7 +13,7 @@ pub struct LinuxPlatform {}
 
 impl Platform for LinuxPlatform {
     fn has_root_perms(&self) -> bool {
-        Uid::effective().is_root()
+        common::has_root_perms()
     }
 
     // TODO: in future could have code to find an unused tun
@@ -28,48 +27,7 @@ impl Platform for LinuxPlatform {
         username: &str,
         dry_run: bool,
     ) -> Result<(), PlatformErr> {
-        match get_user_by_name(username) {
-            None => {
-                return Err(PlatformErr::OsError(format!("user {} not found", username)));
-            }
-            Some(user) => {
-                if dry_run {
-                    println!(
-                        "chown {}:{} {}",
-                        user.uid(),
-                        user.primary_group_id(),
-                        ctrl_path.display()
-                    );
-                } else {
-                    unistd::chown(
-                        ctrl_path,
-                        Some(Uid::from_raw(user.uid())),
-                        Some(Gid::from_raw(user.primary_group_id())),
-                    )
-                    .map_err(|e| PlatformErr::OsError(format!("chown failed: {}", e)))?;
-                }
-            }
-        };
-
-        // Now set perm to 775
-        if dry_run {
-            println!("chmod 775 {}", ctrl_path.display());
-            return Ok(());
-        }
-
-        let dirfd = nix::fcntl::open(
-            ctrl_path,
-            nix::fcntl::OFlag::O_DIRECTORY,
-            nix::sys::stat::Mode::S_IRWXU,
-        )
-        .map_err(|e| PlatformErr::OsError(format!("open failed on {:?}: {}", ctrl_path, e)))?;
-
-        stat::fchmod(
-            dirfd,
-            stat::Mode::S_IRWXU | stat::Mode::S_IRWXG | stat::Mode::S_IROTH | stat::Mode::S_IXOTH,
-        )
-        .map_err(|e| PlatformErr::OsError(format!("fchmod failed: {}", e)))?;
-        Ok(())
+        common::set_control_dir_owner_and_perms(ctrl_path, username, dry_run)
     }
 
     fn drop_privledges(&self, username: &str, dry_run: bool) -> Result<(), PlatformErr> {
@@ -185,12 +143,7 @@ impl Platform for LinuxPlatform {
         Ok(())
     }
 
-    fn exec(&self, mut cmd: Command, dry_run: bool) -> Result<(), PlatformErr> {
-        if dry_run {
-            println!("exec {:?}", cmd);
-            return Ok(());
-        }
-        let err = cmd.exec();
-        Err(PlatformErr::IoError(err))
+    fn exec(&self, cmd: Command, dry_run: bool) -> Result<(), PlatformErr> {
+        common::exec(cmd, dry_run)
     }
 }
